requirements.txt
streamlit
langchain
langchain_google_genai
google-cloud-aiplatform
google-generativeai
dotenv